Refactor the Repositories
