+++ server/src/app/api/activity/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { ActivityRepository } from "@/repositories/repositories";
+import { ActivityRepository } from "@/repositories/activity_Repository_impl";
 import getSession from "@/server_actions/getSession";
 
 const activityRepository = new ActivityRepository();
