+++ server/src/repositories/project_repository_impl.ts
@@ -1,322 +1,10 @@
 import prisma from "@/lib/prisma";
 import {
-  User,
-  Activity,
-  MentorActivity,
-  Report,
-  MentorFeedback,
-  Invitation,
-  Role,
   Project,
-  ProjectAllocation,
 } from "@prisma/client";
 import BaseRepository from "./baseRepository";
-import crypto from "crypto"
 
-/*
 
-      //-- User Repository --//
-
-*/
-
-export class UserRepository extends BaseRepository<User> {
-  constructor() {
-    super(prisma.user);
-  }
-
-  getByEmail(email: string): Promise<User | null> {
-    return this.modelClient.findUnique({
-      where: { email },
-    });
-  }
-
-  async getUserWithActivities(studentId: string) {
-    return this.modelClient.findUnique({
-      where: {
-        id: studentId,
-      },
-      select: {
-        firstName: true,
-        lastName: true,
-        activities: {
-          select: {
-            date: true,
-            timeSpent: true,
-            notes: true,
-            feedback: {
-              select: {
-                status: true,
-                feedbackNotes: true,
-              },
-            },
-          },
-        },
-      },
-    });
-  }
-}
-
-/*
-
-      //-- Activity Repository --//
-
-*/
-
-export class ActivityRepository extends BaseRepository<Activity> {
-  constructor() {
-    super(prisma.activity);
-  }
-
-  async findByStudentId(studentId: string, date?: Date) {
-    return this.modelClient.findMany({
-      where: {
-        studentId,
-        ...(date && { date }),
-      },
-      include: {
-        feedback: {
-          select: {
-            status: true,
-            feedbackNotes: true,
-          },
-        },
-      },
-    });
-  }
-
-  async createActivity(
-    studentId: string,
-    date: Date,
-    timeSpent: number,
-    notes: string,
-  ) {
-    return this.modelClient.create({
-      data: {
-        studentId,
-        date,
-        timeSpent,
-        notes,
-      },
-    });
-  }
-
-  async updateActivity(
-    id: string,
-    studentId: string,
-    data: { timeSpent?: number; notes?: string },
-  ) {
-    return this.modelClient.update({
-      where: {
-        id,
-        studentId,
-      },
-      data,
-    });
-  }
-
-  async findActivityById(id: string) {
-    return this.modelClient.findUnique({
-      where: { id },
-      select: { createdAt: true, studentId: true },
-    });
-  }
-
-  async deleteActivity(id: string) {
-    return this.modelClient.delete({
-      where: { id },
-    });
-  }
-
-  async getStudentFeedbacks(studentId: string, date?: string) {
-    return this.modelClient.findMany({
-      where: {
-        studentId: studentId,
-        ...(date && { date: new Date(date) }),
-      },
-      include: {
-        feedback: {
-          select: {
-            status: true,
-            feedbackNotes: true,
-          },
-        },
-      },
-    });
-  }
-}
-
-/*
-
-      //-- Mentor Activity Repository --//
-
-*/
-
-export class MentorRepository extends BaseRepository<MentorActivity> {
-  constructor() {
-    super(prisma.mentorActivity);
-  }
-
-  async getMentorActivities(mentorId: string, date?: Date) {
-    return this.modelClient.findMany({
-      where: {
-        mentorId,
-        ...(date && { date }),
-      },
-    });
-  }
-
-  async createMentorActivity(data: {
-    mentorId: string;
-    date: Date;
-    workingHours: number;
-    activities: string;
-  }) {
-    return this.modelClient.create({
-      data,
-    });
-  }
-
-  async updateMentorActivity(
-    id: string,
-    mentorId: string,
-    data: {
-      workingHours?: number;
-      activities?: string;
-    },
-  ) {
-    return this.modelClient.update({
-      where: {
-        id,
-        mentorId,
-      },
-      data,
-    });
-  }
-}
-
-/*
-
-      //-- Report Repository --//
-
-*/
-
-export class ReportRepository extends BaseRepository<Report> {
-  constructor() {
-    super(prisma.report);
-  }
-}
-
-/*
-
-      //-- Mentor Feedback Repository --//
-
-*/
-
-export class MentorFeedbackRepository extends BaseRepository<MentorFeedback> {
-  constructor() {
-    super(prisma.mentorFeedback);
-  }
-
-  async getFeedbackByActivityId(activityId: string, date?: Date) {
-    return this.modelClient.findFirst({
-      where: {
-        activityId: String(activityId),
-        activity: {
-          date: date ? new Date(date) : undefined,
-        },
-      },
-    });
-  }
-
-  async upsertFeedback(
-    activityId: string,
-    mentorId: string,
-    review: string,
-    status: string,
-  ) {
-    // Check if the feedback already exists
-    const existingFeedback = await this.modelClient.findFirst({
-      where: {
-        activityId,
-        mentorId,
-      },
-    });
-
-    if (existingFeedback) {
-      // Update the existing feedback
-      return this.modelClient.update({
-        where: {
-          id: existingFeedback.id,
-        },
-        data: {
-          feedbackNotes: review,
-          status,
-        },
-      });
-    } else {
-      // Create new feedback
-      return this.modelClient.create({
-        data: {
-          activityId,
-          mentorId,
-          feedbackNotes: review,
-          status,
-        },
-      });
-    }
-  }
-}
-export class InvitationRepository extends BaseRepository<Invitation> {
-  constructor() {
-    super(prisma.invitation);
-  }
-
-  async createInvite(data: {
-    email: string;
-    role: Role;
-    invitedBy: string; // ID of the super-admin
-    tempPassword: string; // hashed
-  }) {
-    const token = crypto.randomBytes(32).toString("hex"); // secure token
-    const expiresAt = new Date(Date.now() + 3 * 60 * 60 * 1000); // 3 hours
-
-    //to make sure invitedBy is not null
-    if (!data.invitedBy)
-      throw new Error("invitedBy (super-admin ID) is required");
-
-    return this.modelClient.create({
-      data: {
-        email: data.email,
-        role: data.role,
-        token,
-        expiresAt,
-        accepted: false,
-        tempPassword: data.tempPassword,
-        inviter: {
-          connect: { id: data.invitedBy }, // ✅ Connect the relation properly
-        },
-      },
-    });
-  }
-
-  async findValidInvite(email: string, token: string) {
-    return this.modelClient.findFirst({
-      where: {
-        email,
-        token,
-        accepted: false,
-        expiresAt: { gte: new Date() },
-      },
-    });
-  }
-
-  async markAccepted(id: string) {
-    return this.modelClient.update({
-      where: { id },
-      data: { accepted: true },
-    });
-  }
-}
 //-- Project Repository --//
 // Manages projects that organize mentorships by grouping students
 // under mentors within specific domains (software, film, training, research, other)
