+++ server/src/repositories/user_repository_impl.ts
@@ -1,0 +1,49 @@
+import prisma from "@/lib/prisma";
+import { User,} from "@prisma/client";
+import BaseRepository from "./baseRepository";
+
+/*
+
+      //-- User Repository --//
+
+*/
+
+export class UserRepository extends BaseRepository<User> {
+  constructor() {
+    super(prisma.user);
+  }
+
+  getByEmail(email: string): Promise<User | null> {
+    return this.modelClient.findUnique({
+      where: { email },
+    });
+  }
+
+  async getUserWithActivities(studentId: string) {
+    return this.modelClient.findUnique({
+      where: {
+        id: studentId,
+      },
+      select: {
+        firstName: true,
+        lastName: true,
+        activities: {
+          select: {
+            date: true,
+            timeSpent: true,
+            notes: true,
+            feedback: {
+              select: {
+                status: true,
+                feedbackNotes: true,
+              },
+            },
+          },
+        },
+      },
+    });
+  }
+}
+
+
+
