+++ server/test/invitationRepository.test.ts
@@ -1,4 +1,4 @@
-import { InvitationRepository } from "../src/repositories/repositories";
+import { InvitationRepository } from "../src/repositories/user_repository_impl";
 import prisma from "../src/lib/prisma";
 import { Role } from "@prisma/client";
 import crypto from "crypto";
